More data and season filtering for fish ingredients
Fish have much more particular requirements for obtaining than crops and
are not a shipping item in and of themselves (so season data was not
being pulled). This commit brings in the information and displays from
the Fishing page so that filtering fish by season works and information
about catching them is readily available.